Cleaned up homepage. Removed unnecessary shared components.
