--- src/data/siteSummary.json
@@ -1,20 +1,0 @@
-{
-    "featuredPosts": [
-        {
-            "title": "Very Interesting Post",
-            "image": "https://images.unsplash.com/photo-1683009427540-c5bd6a32abf6?ixlib=rb-4.0.3&ixid=MnwxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80",
-            "postDate": "Jan 11, 2021",
-            "readTime": "3 min",
-            "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vel fermentum sem, quis bibendum quam.",
-            "tags": ["something", "something"]
-        },
-        {
-            "title": "Most recent post",
-            "image": "https://images.unsplash.com/photo-1683009427540-c5bd6a32abf6?ixlib=rb-4.0.3&ixid=MnwxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80",
-            "postDate": "May 3, 2023",
-            "readTime": "1 min",
-            "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vel fermentum sem, quis bibendum quam.",
-            "tags": ["blah", "blah"]
-        }
-    ]
-}
